fix table headers menu and change menus for more, query search
