+++ frontend/public/templates/work_packages/work_packages_table.html
@@ -30,7 +30,8 @@
               position-relative-to=".sort-header-outer"
               locals="columns, column"
               sortable="true"
-              query="query"/>
+              query="query">
+          </th>
 
           <th sort-header ng-repeat="column in columns"
                           has-dropdown-menu
